In progress, gettting close

+++ SGP40.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 
 int SGP40_start();
-int SGP40_read(float temp, float humidity, float* VOC);
+int SGP40_read(float temp, float humidity, int32_t* VOC);
